The great renaming: Crow => Cro.
First and foremost, there's already a somewhat well publicised web
framework called Crow, and it's hard to argue that this is "a service
framework not a web framework" to make it "ok"ish. Best not to have
the conflict in the first place.

Second, because the name was a pun on `my Crow::Service $foo` reading
aloud as "microservice", `my Cro::Service $foo` is even closer. "Cro,
as in micro" works well.

Third, because I could register `cro.services`, while `crow.services`
and all the others were taken. This also means `mi.cro.services` is a
viable domain to stick on slides etc. Which is just cute.

Fourth, because after buying a .services domain I can't afford anyone
who can draw me a crow to make the obvious logo anyway - but now it's
called "Cro" I can probably craft an acceptable logo myself. :P